meta-feedback-actor: Use float for position and anchor
To be correct with fractional scaling. Furthermore, we currently
use it only with ClutterPoint values, which are floats already.

https://gitlab.gnome.org/GNOME/mutter/merge_requests/684

+++ src/compositor/meta-dnd-actor.c
@@ -206,7 +206,7 @@ meta_dnd_actor_drag_finish (MetaDnDActor *self,
 
       if (CLUTTER_ACTOR_IS_VISIBLE (self->drag_origin))
         {
-          int anchor_x, anchor_y;
+          float anchor_x, anchor_y;
           ClutterPoint dest;
 
           clutter_actor_get_transformed_position (self->drag_origin,
